requirements.txt
# Web
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sse-starlette==2.1.3

# Settings & validation
# pydantic pinned <2.10 because aiogram 3.15 requires it
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.14.0
pgvector==0.3.6

# Auth
passlib[bcrypt]==1.7.4
pyjwt==2.10.1

# Queue
arq==0.26.3
redis==5.2.0

# AI
openai==1.55.3
tiktoken==0.8.0

# Transcription
youtube-transcript-api==0.6.3
yt-dlp==2024.11.18
pydub==0.25.1

# Storage
boto3==1.35.79

# HTTP client
httpx==0.28.0

# Telegram (lightweight, just for posting)
aiogram==3.15.0

# Utils
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
